+++ EssentialApp/EssentialApp/SceneDelegate.swift
@@ -12,6 +12,9 @@ import CoreData
 
 class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     var window: UIWindow?
+    let localStoreURL = NSPersistentContainer
+        .defaultDirectoryURL()
+        .appendingPathExtension("feed-store.sqlite")
 
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let _ = (scene as? UIWindowScene) else { return }
@@ -21,16 +24,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         let remoteClient = makeRemoteClient()
         let remoteFeedLoader = RemoteFeedLoader(url: remoteURL, client: remoteClient)
         let remoteImageLoader = RemoteFeedImageDataLoader(client: remoteClient)
-
-        let localStoreURL = NSPersistentContainer
-            .defaultDirectoryURL()
-            .appendingPathExtension("feed-store.sqlite")
-
-        #if DEBUG
-        if CommandLine.arguments.contains("-reset") {
-            try? FileManager.default.removeItem(at: localStoreURL )
-        }
-        #endif
 
         let localStore = try! CoreDataFeedStore(storeURL: localStoreURL)
         let localFeedLoader = LocalFeedLoader(store: localStore, currentDate: Date.init)
@@ -54,31 +47,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
         )
     }
     
-    private func makeRemoteClient() -> HTTPClient {
-        #if DEBUG
-        if UserDefaults.standard.string(forKey: "connectivity") == "offline" {
-            return AlwaysFailingClient()
-        }
-        #endif
-
-        return URLSessionHTTPClient(session: URLSession(configuration: .ephemeral))
+    func makeRemoteClient() -> HTTPClient {
+        URLSessionHTTPClient(session: URLSession(configuration: .ephemeral))
     }
 }
-
-#if DEBUG
-private class AlwaysFailingClient: HTTPClient {
-    private class Task: HTTPClientTask {
-        func cancel() {}
-    }
-
-    func get(
-        from url: URL,
-        completion: @escaping (HTTPClient.Result) -> Void
-    ) -> any EssentialFeed.HTTPClientTask {
-        completion(.failure(NSError(domain: "offline", code: 0)))
-
-        return Task()
-    }
-}
-#endif
-
